# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ticket Sell API",
    "description": "API for selling tickets. All responses follow the envelope pattern: `{ success, data, message, code }`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Default server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "description": "Response payload. Shape varies by endpoint."
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable success code in SCREAMING_SNAKE_CASE",
            "example": "USER_CREATED"
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "code"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "array",
            "items": {},
            "example": []
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials, verify the email or password"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code in SCREAMING_SNAKE_CASE",
            "example": "INVALID_CREDENTIALS"
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "code"
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid Bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "data": [],
              "message": "Authentication required. Please provide a valid Bearer token.",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/tickets/{id}": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Get a ticket",
        "description": "Returns a ticket product. `quantity` is the total stock.",
        "operationId": "ticketsGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Ticket cuid",
            "schema": {
              "type": "string",
              "example": "clx0ticket0001abcdefghijk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "clx0ticket0001abcdefghijk",
                    "title": "Rock Festival 2026",
                    "description": "Main stage, general admission",
                    "price": 150,
                    "quantity": 100,
                    "createdAt": "2026-09-24T12:00:00.000Z",
                    "updatedAt": "2026-09-24T12:00:00.000Z"
                  },
                  "message": "Ticket fetched successfully.",
                  "code": "TICKET_FETCHED"
                }
              }
            }
          },
          "404": {
            "description": "No ticket exists with the given id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "data": [],
                  "message": "Ticket not found.",
                  "code": "TICKET_NOT_FOUND"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{id}/reservation": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Get the authenticated user's reservation",
        "description": "Returns the open reservation of the user identified by the Bearer token for this ticket. A completed purchase is not a reservation.",
        "operationId": "ticketsGetReservation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Ticket cuid",
            "schema": {
              "type": "string",
              "example": "clx0ticket0001abcdefghijk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "clx0reserv0001abcdefghijk",
                    "ticketId": "clx0ticket0001abcdefghijk",
                    "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "quantity": 2,
                    "status": "RESERVED",
                    "createdAt": "2026-09-24T12:30:00.000Z",
                    "ticket": {
                      "id": "clx0ticket0001abcdefghijk",
                      "title": "Rock Festival 2026",
                      "description": "Main stage, general admission",
                      "price": 150,
                      "quantity": 100,
                      "createdAt": "2026-09-24T12:00:00.000Z",
                      "updatedAt": "2026-09-24T12:00:00.000Z"
                    }
                  },
                  "message": "Ticket reservation fetched successfully.",
                  "code": "TICKET_RESERVATION_FETCHED"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "The ticket does not exist, or the authenticated user has no open reservation for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "ticketNotFound": {
                    "summary": "Ticket does not exist",
                    "value": {
                      "success": false,
                      "data": [],
                      "message": "Ticket not found.",
                      "code": "TICKET_NOT_FOUND"
                    }
                  },
                  "reservationNotFound": {
                    "summary": "No open reservation for this user",
                    "value": {
                      "success": false,
                      "data": [],
                      "message": "Reservation not found.",
                      "code": "RESERVATION_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}